feat: add support for standalone Home Manager configurations
This commit adds support for standalone Home Manager configurations
defined in a 'users' subdirectory of any host.

Defining Home Manager configurations per-host gives users the
flexibility to either write a fully custom config or delegate to a
shared module using the established 'modules/home' directory.

Configuration definitions can take two possible forms:
1. hosts/<hostname>/users/<username>.nix
2. hosts/<hostname>/users/<username>/home-configuration.nix

If the host defines a NixOS or nix-darwin configuration, and the input
'home-manager' exists, then the Home Manager module for that host will
be imported and all defined users will be added automatically.
Additionally, the extraSpecialArgs is modified to include 'flake',
'inputs', and 'perSystem'. An error is thrown if the 'home-manager'
input does not exist.

Standalone configurations also have the 'home.username' and
'home.homeDirectories' options set using 'lib.mkDefault'. This allows it
to be overridden by a user.

If a 'users' directory exists for a given host, that host directory no
longer needs to contain a host configuration file.