feat: populate template description using flake (#63)
This commit fixes a FIXME, using a more meaningful description for
templates. Nix flakes allow a description field which, if populated,
makes sense to use as a description for the template too.

Template descriptions can easily be written in such a way that the
description works to both describe the flake and exist in the user's
repository until changed. For example, all the templates in
github:nixos/templates have a description field in their flakes that
would also make sense to see in 'nix flake show' output.

While a better description could probably be provided manually, this is
an elegant way to infer a more meaningful description without having to
add any extra files or non-standard fields.