feat: home configurations in top-level users dir
Blueprint now supports defining user configurations in the top-level
'users' directory, and makes this a first-class way to share user
configurations without exposing them in the public interface of the
flake.

Users are created in 'users/<username>/home-configuration.nix' (though
this is easy to change in the future).

Existing home configurations can access all the generic configurations
through the new 'users' argument. For example, a user defined at
'users/baz/home-configuration.nix' could be imported and used as a base
by 'hosts/foobar/users/baz.nix' (`imports = [ users.baz ];`) OR as a
standalone configuration without a hostname.

Editorial: new users should still be encouraged to start creating their
configuration per-machine to reduce the complexity of their config, and
only move to a shared config when they _need_ to share it between
multiple devices.