Make darwin/home/nixosModules not be set when no modules are present
Right now, if you use blueprint in a flake, the outputs
darwin/home/nixosModules will be set to a empty attribute if there are
no modules, this can be easily fixed with `lib.mkIf`. This error also
exists with darwin/system/nixosConfigurations, and using `lib.mkIf` will
make those lines quite long, I don't have a solution for that currently.