fix: don't reevaluate nixpkgs in nixosConfigurations
Before this fix, `pkgs` was reevaluated in NixOS configurations. If any `nixpkgs.config` or `nixpkgs.overlays` was set, it got lost. Besides, this caused a performance penalty.

Besides, it was impossible to redefine `nixpkgs.pkgs` downstream because `perSystemModule` depended on `pkgs`. Any redefinition caused an infinite loop. Now it depends on `nixpkgs.hostPlatform`, which must be set in each machine configuration, so the loop is gone.

@moduon MT-9339