Pass lib to lib/default.nix
When blueprint imports `lib/default.nix`, it only passes inputs and
flake. Nixpkgs lib will probaly be used in the function so we can pass
it, since `inputs.nixpkgs.lib` doesn't depend on system and the
`nixpkgs` input will always be available because the inputs of blueprint
and the guest are combined.